refactor(article,series): integrate series api modules to article api modules
