+++ services/article/api/exception/src/main/java/nettee/blolet/article/exception/SeriesArticleErrorCode.java
@@ -1,4 +1,4 @@
-package nettee.series.article.exception;
+package nettee.blolet.article.exception;
 
 import nettee.common.ErrorCode;
 import org.springframework.http.HttpStatus;
