+++ services/article/api/exception/src/main/java/nettee/blolet/article/exception/SeriesArticleException.java
@@ -1,4 +1,4 @@
-package nettee.series.article.exception;
+package nettee.blolet.article.exception;
 
 import nettee.common.CustomException;
