+++ services/series/application/src/main/java/nettee/series/application/usecase/SeriesVisitUseCase.java
@@ -1,6 +1,6 @@
 package nettee.series.application.usecase;
 
-import nettee.series.readmodel.SeriesQueryModels.SeriesDetail;
+import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 
 public interface SeriesVisitUseCase {
     SeriesDetail findDetailForPublic(String seriesId);
