+++ services/series/application/src/main/java/nettee/series/article/application/service/SeriesArticleCommandService.java
@@ -1,16 +1,16 @@
 package nettee.series.article.application.service;
 
 import lombok.RequiredArgsConstructor;
+import nettee.blolet.article.domain.SeriesArticle;
 import nettee.series.article.application.port.SeriesArticleCommandRepositoryPort;
 import nettee.series.article.application.usecase.SeriesArticleCreateUseCase;
 import nettee.series.article.application.usecase.SeriesArticleDeleteUseCase;
 import nettee.series.article.application.usecase.SeriesArticleUpdateUseCase;
-import nettee.series.article.domain.SeriesArticle;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static nettee.series.article.exception.SeriesArticleErrorCode.SERIES_ARTICLE_NOT_FOUND;
+import static nettee.blolet.article.exception.SeriesArticleErrorCode.SERIES_ARTICLE_NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
