+++ services/series/application/src/main/java/nettee/series/article/application/usecase/SeriesArticleReadUseCase.java
@@ -1,6 +1,6 @@
 package nettee.series.article.application.usecase;
 
-import nettee.series.article.readmodel.SeriesArticleQueryModels.SeriesArticleSummary;
+import nettee.blolet.article.readmodel.SeriesArticleQueryModels.SeriesArticleSummary;
 
 import java.util.List;
