+++ services/series/application/src/main/java/nettee/series/article/application/usecase/SeriesArticleUpdateUseCase.java
@@ -1,6 +1,6 @@
 package nettee.series.article.application.usecase;
 
-import nettee.series.article.domain.SeriesArticle;
+import nettee.blolet.article.domain.SeriesArticle;
 
 public interface SeriesArticleUpdateUseCase {
