+++ services/series/driven/rdb/src/main/java/nettee/series/article/driven/rdb/persistence/SeriesArticleQueryAdapter.java
@@ -1,9 +1,9 @@
 package nettee.series.article.driven.rdb.persistence;
 
+import nettee.blolet.article.readmodel.SeriesArticleQueryModels.SeriesArticleSummary;
 import nettee.series.article.application.port.SeriesArticleQueryRepositoryPort;
 import nettee.series.article.driven.rdb.entity.SeriesArticleEntity;
 import nettee.series.article.driven.rdb.persistence.mapper.SeriesArticleEntityMapper;
-import nettee.series.article.readmodel.SeriesArticleQueryModels.SeriesArticleSummary;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 import org.springframework.stereotype.Repository;
