+++ services/series/driving/web-mvc/src/main/java/nettee/series/driving/web/SeriesQueryApi.java
@@ -6,13 +6,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 import nettee.blolet.jwt.filter.annotation.AuthUser;
 import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
 import nettee.series.application.usecase.SeriesReadUseCase;
 import nettee.series.application.usecase.SeriesVisitUseCase;
 import nettee.series.driving.web.dto.SeriesQueryDto.SeriesDetailResponse;
 import nettee.series.driving.web.dto.SeriesQueryDto.SeriesSummaryResponse;
-import nettee.series.readmodel.SeriesQueryModels.SeriesDetail;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
