+++ services/series/driving/web-mvc/src/main/java/nettee/series/driving/web/dto/SeriesCommandDto.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
+import nettee.blolet.article.domain.Series;
 import nettee.series.article.driving.web.dto.SeriesArticleCommandDto.SeriesArticleCreateCommand;
-import nettee.series.domain.Series;
 
 import java.util.List;
