+++ services/series/driving/web-mvc/src/main/java/nettee/series/driving/web/dto/SeriesQueryDto.java
@@ -2,8 +2,8 @@ package nettee.series.driving.web.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
-import nettee.series.readmodel.SeriesQueryModels.SeriesDetail;
-import nettee.series.readmodel.SeriesQueryModels.SeriesSummary;
+import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
+import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesSummary;
 
 import java.util.List;
